move validation logic in to wasm-contract
