+++ Cargo.toml
@@ -44,7 +44,6 @@ zstd = "0.4"
 
 [dev-dependencies]
 tempdir = "0.3"
-wabt = "0.7"
 
 [workspace]
 members = [
